+++ CHANGELOG.md
@@ -4,5 +4,7 @@ What's new in Forest X Reborn 1.3.0?<br />
 - Silver tools are no longer enchanted when crafting
 - Added grapes and grape vine
 - Added a quiver
+- Added ash block
+- Completely redesigned volcano biome
 - Made some improvements with inventory tabs
 - [Bugfix]